requirements.txt
uvicorn
fastapi
pdfplumber
nltk
openai
psycopg2-binary
pgvector
langdetect
python-dotenv
requests
pandas
scikit-learn
pdfplumber
langdetect
psycopg2-binary
uvicorn
fastapi
openai
python-dotenv

